Fix relabel issue

+++ cpp/oneapi/dal/algo/triangle_counting/backend/cpu/vertex_ranking_default_kernel_avx512.hpp
@@ -801,7 +801,7 @@ DAAL_FORCEINLINE std::int64_t triangle_counting_global_scalar_avx512(
     std::int64_t total_s = oneapi::dal::detail::parallel_reduce_int32_int64_t(
         vertex_count,
         (std::int64_t)0,
-        [&](std::int64_t begin_u, std::int64_t end_u, std::int64_t tc_u) -> std::int64_t {
+        [&](std::int32_t begin_u, std::int32_t end_u, std::int64_t tc_u) -> std::int64_t {
             for (auto u = begin_u; u != end_u; ++u) {
                 for (auto v_ = vertex_neighbors + edge_offsets[u];
                      v_ != vertex_neighbors + edge_offsets[u + 1];
@@ -845,7 +845,7 @@ DAAL_FORCEINLINE std::int64_t triangle_counting_global_vector_avx512(
     std::int64_t total_s = oneapi::dal::detail::parallel_reduce_int32_int64_t_simple(
         vertex_count,
         (std::int64_t)0,
-        [&](std::int64_t begin_u, std::int64_t end_u, std::int64_t tc_u) -> std::int64_t {
+        [&](std::int32_t begin_u, std::int32_t end_u, std::int64_t tc_u) -> std::int64_t {
             for (auto u = begin_u; u != end_u; ++u) {
                 if (degrees[u] < 2) {
                     continue;
@@ -902,8 +902,8 @@ DAAL_FORCEINLINE std::int64_t triangle_counting_global_vector_relabel_avx512(
     std::int64_t total_s = oneapi::dal::detail::parallel_reduce_int32_int64_t_simple(
         vertex_count,
         (std::int64_t)0,
-        [&](std::int64_t begin_u, std::int64_t end_u, std::int64_t tc_u) -> std::int64_t {
-            for (auto u = begin_u; u != end_u; ++u) {
+        [&](std::int32_t begin_u, std::int32_t end_u, std::int64_t tc_u) -> std::int64_t {
+            for (auto u = begin_u; u != end_u; ++u) {  
                 if (degrees[u] < 2) {
                     continue;
                 }
